cast chars to int when we want to print integers so we get a number instead of a character
--HG--
extra : convert_revision : 7bfa88ba23ad057b751eb01a80416d9f72cfe81a

+++ src/base/cprintf_formats.hh
@@ -288,13 +288,13 @@ format_integer(std::ostream &out, const T &data, Format &fmt)
 { _format_integer(out, data, fmt); }
 inline void
 format_integer(std::ostream &out, char data, Format &fmt)
-{ _format_integer(out, data, fmt); }
+{ _format_integer(out, (int)data, fmt); }
 inline void
 format_integer(std::ostream &out, unsigned char data, Format &fmt)
-{ _format_integer(out, data, fmt); }
+{ _format_integer(out, (int)data, fmt); }
 inline void
 format_integer(std::ostream &out, signed char data, Format &fmt)
-{ _format_integer(out, data, fmt); }
+{ _format_integer(out, (int)data, fmt); }
 #if 0
 inline void
 format_integer(std::ostream &out, short data, Format &fmt)
